Float over Double

+++ app/src/main/java/com/ps/simplepapersoccer/Helpers/MathHelper.kt
@@ -5,9 +5,9 @@ import kotlin.math.abs
 import kotlin.math.sqrt
 
 object MathHelper {
-    fun euclideanDistance(point: TwoDimensionalPoint, otherPoint: TwoDimensionalPoint): Double {
+    fun euclideanDistance(point: TwoDimensionalPoint, otherPoint: TwoDimensionalPoint): Float {
         val yCord = abs(point.y - otherPoint.y)
         val xCord = abs(point.x - otherPoint.x)
-        return sqrt((yCord * yCord + xCord * xCord).toDouble())
+        return sqrt((yCord * yCord + xCord * xCord).toFloat())
     }
 }
